This commit consists mostly of the addition of a test program of splashdown location estimation.
In addition, some initial implementations of event detectors have been
added so that liftoff, apogee, and landing are now functional.  A
burnout detector exists but has not been implemented.

The splashdown location estimation program consists of a simple 3DOF
rocket flight simulation with noisy wind, noisy GPS readings, and the
potential (currently unimplemented) for noisy IMU readings.  This allows
for the testing of GPS tracking and prediction algorithms.  There is
currently no data processing in the estimation program, so it saves all
the simulation data to a CSV file for later data processing /
visualization.